create cart

+++ packages/medusa/src/models/region.ts
@@ -34,7 +34,7 @@ export class Region {
   @JoinColumn({ name: "currency_code", referencedColumnName: "code" })
   currency: Currency
 
-  @Column({ type: "int" })
+  @Column({ type: "decimal" })
   tax_rate: number
 
   @Column({ nullable: true })
